backend-api/requirements.txt
pygame
kafka-python
pyspark

# ✨ Flask API용 패키지
flask
mysql-connector-python

